refactor: unify locale resolution across admin SSR, RPC, and public routes
Two parallel resolvers used to live side by side: the public/RPC path used a 3-tier chain
(override → user.meta.locale → site default) and the admin SSR used a 4-tier chain (?lang= →
user.meta → cookie → Accept-Language → default). The split was deliberate — keep cookies and
headers off public routes so CDN cache keys don't fragment — but it left a real hole: the
pre-auth locale-cookie path was scoped to `/_plumix/admin/`, so `POST /_plumix/auth/magic-link/
request` never saw it. The dropdown pick on /login localised the page but not the email.

Collapse into one `resolveLocale` with 6 tiers (override → ?lang= → user.meta.locale → cookie →
Accept-Language → site default), widen the cookie path to `/_plumix/` so it reaches admin shell,
auth POSTs, and the RPC pipeline uniformly, and path-gate the Accept-Language tier to
`/_plumix/*` so the public-route cache stays URL-keyed. Net -88 LOC.

End result: a DE user picks DE on /login → cookie + ?lang=de → admin shell SSR renders DE →
submits magic-link form → POST under /_plumix/ carries the cookie → `ctx.locale` reads it →
magic-link email goes out in DE.

`runtime/admin-shell.ts` keeps just `rewriteAdminShellLangDir`; its locale resolution moves to
`i18n/resolve-locale.ts`. The Accept-Language matcher is extracted to `i18n/accept-language.ts`.
TDD per [[tdd-vertical-slicing]] — 5 RED→GREEN cycles, one tier at a time, plus an explicit
3-way precedence test (?lang= over user.meta AND cookie). All 1789 core + 519 admin tests pass;
typecheck, lint, format, build all green.
