+++ packages/core/src/i18n/cookie.test.ts
@@ -11,13 +11,13 @@ import { buildLocaleCookie } from "./cookie.js";
 describe("buildLocaleCookie", () => {
   test("serializes the full cookie attribute set without Secure", () => {
     expect(buildLocaleCookie("uk", false)).toBe(
-      "plumix_locale=uk; Path=/_plumix/admin/; Max-Age=31536000; SameSite=Lax",
+      "plumix_locale=uk; Path=/_plumix/; Max-Age=31536000; SameSite=Lax",
     );
   });
 
   test("appends Secure when called over HTTPS", () => {
     expect(buildLocaleCookie("uk", true)).toBe(
-      "plumix_locale=uk; Path=/_plumix/admin/; Max-Age=31536000; SameSite=Lax; Secure",
+      "plumix_locale=uk; Path=/_plumix/; Max-Age=31536000; SameSite=Lax; Secure",
     );
   });
 });
