+++ packages/core/src/i18n/resolve-locale.ts
@@ -1,5 +1,8 @@
 import type { AuthenticatedUser } from "../context/app.js";
 import type { ResolvedI18n, ResolvedLocale } from "./locale-registry.js";
+import { readSessionCookie } from "../auth/cookies.js";
+import { matchAcceptLanguage } from "./accept-language.js";
+import { ADMIN_LOCALE_COOKIE } from "./cookie.js";
 import { findEnabledLocale } from "./locale-registry.js";
 
 interface ResolveLocaleArgs {
@@ -8,22 +11,36 @@ interface ResolveLocaleArgs {
   readonly i18n: ResolvedI18n;
 }
 
-// WP `determine_locale()` parity: operator override → user.meta.locale →
-// site default. No cookie, no Accept-Language — those fragment public cache
-// keys; layer them back via the override slot if needed.
+// Single source of truth for admin SSR, RPC, and public-route surfaces.
+// The override now fires on admin SSR too (pre-merge admin-shell skipped
+// it); operators who want admin to ignore the override should narrow it
+// on the request path themselves.
 export function resolveLocale({
   request,
   user,
   i18n,
 }: ResolveLocaleArgs): ResolvedLocale {
+  const resolveCode = (
+    code: string | null | undefined,
+  ): ResolvedLocale | null =>
+    code ? (findEnabledLocale(i18n, code) ?? null) : null;
+
   const override = i18n.resolveLocale?.(request, user);
-  if (override) {
-    const match = findEnabledLocale(i18n, override.code);
-    if (match) return match;
-  }
-  if (typeof user?.meta.locale === "string") {
-    const match = findEnabledLocale(i18n, user.meta.locale);
-    if (match) return match;
-  }
-  return i18n.defaultLocale;
+  const url = new URL(request.url);
+  // Path-gate Accept-Language to `/_plumix/*` so public-route HTML stays
+  // identical per URL — CDN cache keys must not fragment by browser locale.
+  // The cookie is already path-gated by the browser via its `Path=/_plumix/`
+  // attribute; we don't re-check the URL for it.
+  const onInternalPath = url.pathname.startsWith("/_plumix/");
+
+  return (
+    resolveCode(override?.code) ??
+    resolveCode(url.searchParams.get("lang")) ??
+    resolveCode(
+      typeof user?.meta.locale === "string" ? user.meta.locale : null,
+    ) ??
+    resolveCode(readSessionCookie(request, ADMIN_LOCALE_COOKIE)) ??
+    (onInternalPath ? matchAcceptLanguage(request, i18n) : null) ??
+    i18n.defaultLocale
+  );
 }
